factor out pending types into own module and move some functions move Pending types into pending module move calculation methods from reddit_comment to pending rename factorial to calculated move limits to pending reorder some items in modules
