+++ src/main.rs
@@ -9,9 +9,10 @@ use std::time::SystemTime;
 use time::OffsetDateTime;
 use tokio::time::{sleep, Duration};
 
-mod factorial;
+mod calculated;
 mod influxdb;
 mod math;
+mod pending;
 mod reddit_api;
 pub(crate) mod reddit_comment;
